+++ ui/components/status/StatusBar.vue
@@ -2,7 +2,15 @@
   <div class="div__status-bar">
     <span class="span__status-bar-text">{{ status_label }}: {{ alert_txt }}</span>
     <span>
-      <b-modal id="error-catch" size="sm" hide-footer hide-header hide-header-close :static="true">
+      <b-modal
+        id="error-catch"
+        size="sm"
+        hide-footer
+        hide-header
+        hide-header-close
+        :static="true"
+        :no-close-on-backdrop="true"
+      >
         <ErrorCatchWidget :log_filepath="log_path" @ok-clicked="close_modals_by_id(['error-catch'])" />
       </b-modal>
       <b-modal